Got Carryover-12 working and fixed a couple of compiler warnings.
git-svn-id: https://groke.otago.ac.nz/ant/ant@206 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/compress.h
@@ -8,7 +8,7 @@
 
 #include "fundamental_types.h"
 
-typedef uint32_t ANT_compressable_integer;
+typedef uint64_t ANT_compressable_integer;
 
 /*
 	class ANT_COMPRESS
